Antisway bugfix in internal message queue

+++ ssabox/lib/ssabox/src/ssabox_as_phasor_methods.h
@@ -241,7 +241,7 @@
   (object->mparams[3]) // Default 30. Hold the current message this number of
                        // cycles.
 #define AS_OBJ_MESSAGEQP(i)                                                    \
-  ((AS_messageQ*)(unsigned long int)object->messageQ[(i)])
+  ((AS_messageQ*)object->messageQ[(i)])
 #define AS_OBJ_MESSAGEQPP(i) ((AS_messageQ**)&object->messageQ[(i)])
 #define AS_ADDMESSAGE(type, charpp, messageArgs...)                            \
   {                                                                            \
